+++ src/core/login/login-delete.ts
@@ -12,7 +12,7 @@ export async function deleteLogin(
   login: LoginTree
 ): Promise<void> {
   const { stashTree } = getStashById(ai, login.loginId)
-  const { deviceDescription } = ai.props.state.login
+  const { deviceDescription } = ai.props.state.login.deviceInfo
 
   const request = makeAuthJson(stashTree, login)
   if (deviceDescription != null) request.deviceDescription = deviceDescription