+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Add `appVersion`, `osType`, and `osVersion` to `EdgeContextOptions` for v2 coreRollup endpoint support.
+
 ## 2.38.3 (2026-01-06)
 
 - added: Added `change2.edge.app` server to default change server list.